inference-api/requirements.txt
fastapi
uvicorn
mlflow
python-dotenv
pandas
scikit-learn